+++ src/Helpers/Commands.ts
@@ -1,20 +1,24 @@
 export class Commands {
     public static PortalList() {
-        return 'pac paportal list';
+        return 'pac powerpages list';
     }
 
     public static DownloadPortal(localPath: string, websiteId: string, overwritePortal: string | undefined, includedEntities?: string, excludedEntities?: string) {
         const overwriteText = overwritePortal === 'Yes' ? `-o true` : '';
         const includeEntities = includedEntities ? `-ie "${includedEntities}"`: '';
         const excludeEntities = excludedEntities ? `-xe "${excludedEntities}"`: '';
-        return `pac paportal download -p "${localPath}" -id ${websiteId} ${overwriteText} ${includeEntities} ${excludeEntities}`;
+        return `pac powerpages download -p "${localPath}" -id ${websiteId} ${overwriteText} ${includeEntities} ${excludeEntities}`;
     }
 
     public static UploadPortal(localPath: string, deploymentProfile: string | undefined) {
         const command = deploymentProfile
-            ? `pac paportal upload -p "${localPath}" --deploymentProfile "${deploymentProfile}"`
-            : `pac paportal upload -p "${localPath}"`;
+            ? `pac powerpages upload -p "${localPath}" --deploymentProfile "${deploymentProfile}"`
+            : `pac powerpages upload -p "${localPath}"`;
         return command;
+    }
+
+    public static BootstrapMigrate(localPath: string){
+        return `pac powerpages bootstrap-migrate -p "${localPath}"`;
     }
 
     public static AuthList() {